+++ requirements_dev.txt
@@ -1,3 +1,4 @@
+-r requirements.txt
 black==22.10.0
 flake8==5.0.4
 isort==5.10.1
