+++ tdmgr/schemas/discovery.py
