+++ tdmgr/schemas/result.py
