+++ tdmgr/schemas/status.py
